requirements.txt
fastapi
uvicorn
google-generativeai
beautifulsoup4
httpx
python-dotenv
supabase
